Move ContentManagerExtensions to Rampastring.XNAUI.Extensions namespace
