+++ FontManagement/FontManager.cs
@@ -7,6 +7,7 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
 using Rampastring.Tools;
+using Rampastring.XNAUI.Extensions;
 
 namespace Rampastring.XNAUI.FontManagement;
